Refactor filter expeditions
