+++ components/home/homeContainer.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Container } from 'semantic-ui-react'
 import Carousel from './carousel.jsx'
-import FilterPopular from './filterPopular.jsx'
+import PopularExpeditions from './popularExpeditions.jsx'
 import { inject, observer } from 'mobx-react'
 import ExpeditionsMap from '../shared/maps/ExpeditionsMap'
 import FeaturedExpedition from './FeaturedExpedition'
